Edit API edited CORS AND if about color

+++ filters/allFormat.js
@@ -20,7 +20,7 @@ export default () => {
     if (!value) return "";
     const options = {
       year: "numeric",
-      month: "numeric",
+      month: "long",
       day: "numeric",
       hour: "numeric",
       minute: "numeric",
